Create containedspace.py
Simulation of an enclosed space (ContainedSpace) that can have heat sources dynamically added and removed. Each heat source can have a different (static) heat output (in Watts). The space has a minimum temperature that it will steadily cool down to, if there are no heat sources. See the documentation string and comments of class ContainedSpace for more info. See test code for examples how to use the class in a simulation.